+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## v0.1.0-rc-10
 
-- Added `stats` as a required method
+- Added `stats` and `unlink` as a required method
 - Added `etag` as arg to stream constructor
 - `S3Stream.iter_dir()` will now yield streams with etag
 - Change signature of `AbcStream.iter_dir_()` to return `Iterable[StreamInfo]` instead